pyproject.toml
[tool.poetry]
name = "stoel"
version = "0.1.0"
description = "Bandwidth measurement tool."
authors = ["Jan Willems <jw@elevenbits.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
speedtest-cli = "^2.1.3"
pydantic = "^1.9.0"
SQLAlchemy = "^1.4.32"
psycopg2 = "^2.9.3"
fastapi = "^0.75.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.0"
pre-commit = "^2.17.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
